refactored code. added defines to make acpi writes more explicit and split the main.c files in its different components. Namely the acpi access part, the functions tasked with handling the powermodes and finally the functions tasked with handling the different settings related to the battery.
