+++ src/main.c
@@ -1,276 +1,6 @@
 #include "power_manager.h"
-enum e_powerstate {IC = 0, EP, BS};
 
-static const char	*query_acpi_info(const char *r_value)
-{
-  int fd;
-  FILE *file;
-  char *value;
-  int ret;
-
-  value = calloc(51, sizeof(char));
-  // Open /proc/acpi/call for writing with elevated privileges
-  file = fopen("/proc/acpi/call", "w");
-  if (file == NULL) {
-      perror("Error opening /proc/acpi/call");
-      return (NULL);
-  }
-
-  // Write the value to the file
-  if (fprintf(file, "%s", r_value) < 0) {
-      perror("Error writing to /proc/acpi/call");
-      fclose(file);
-      return (NULL);
-  }
-
-  // Close the file
-  fclose(file);
-  fd = open("/proc/acpi/call", 0, O_RDONLY);
-  if (fd == -1)
-  {
-	perror("error opening /proc/acpi/call");
-	return (NULL);
-  }
-  ret = read(fd, value, 50);
-  if (!ret)
-  {
-	  perror ("error reading from /proc/acpi/call");
-	  close(fd);
-	  return (NULL);
-  }
-  return (value);
-}
-
-static const char	*query_PwrMode_info(void)
-{
-  const char *value = query_acpi_info("\\_SB.PCI0.LPC0.EC0.SPMO");
-  if (!value)
-	  return ("Cannot determine Power Mode");
-  if (!strncmp("0x0", value, 4))
-  {
-	  free((char *)value);
-	  return ("Intelligent Cooling");
-  }
-  else if (!strncmp("0x1", value, 4))
-  {
-	  free((char *)value);
-	  return ("Extreme Performance");
-  }
-  else if (!strncmp("0x2", value, 4))
-  {
-	  free((char *)value);
-	  return ("Battery Saving");
-  }
-  else
-	  return ("Cannot determine Power Mode");
-}
-
-void update_PwrMode_text(GtkLabel *label)
-{
-    gtk_label_set_text(label, query_PwrMode_info());
-}
-
-static void set_IC(GtkWidget *widget, gpointer data)
-{
-  GObject *label = (GObject *)data;
-  g_print ("Intelligent Cooling Mode Activated\n");
-  FILE *file;
-  const char *value = "\\_SB.PCI0.LPC0.EC0.VPC0.DYTC 0x000FB001";
-  
-  // Open /proc/acpi/call for writing with elevated privileges
-  file = fopen("/proc/acpi/call", "w");
-  if (file == NULL) {
-      perror("Error opening /proc/acpi/call");
-      return;
-  }
-
-  // Write the value to the file
-  if (fprintf(file, "%s", value) < 0) {
-      perror("Error writing to /proc/acpi/call");
-      fclose(file);
-      return;
-  }
-
-  // Close the file
-  fclose(file);
-  update_PwrMode_text(GTK_LABEL(label));
-  (void)widget;
-  (void)data;
-}
-
-static void set_EP(GtkWidget *widget, gpointer   data)
-{
-  GObject *label = (GObject *)data;
-  g_print ("Extreme Performance Mode Activated\n");
-  FILE *file;
-  const char *value = "\\_SB.PCI0.LPC0.EC0.VPC0.DYTC 0x0012B001";
-  
-  // Open /proc/acpi/call for writing with elevated privileges
-  file = fopen("/proc/acpi/call", "w");
-  if (file == NULL) {
-      perror("Error opening /proc/acpi/call");
-      return;
-  }
-
-  // Write the value to the file
-  if (fprintf(file, "%s", value) < 0) {
-      perror("Error writing to /proc/acpi/call");
-      fclose(file);
-      return;
-  }
-
-  // Close the file
-  fclose(file);
-  update_PwrMode_text(GTK_LABEL(label));
-  (void)widget;
-  (void)data;
-}
-
-static void set_BS(GtkWidget *widget, gpointer   data)
-{
-  GObject *label = (GObject *)data;
-  g_print ("Battery Saving mode activated\n");
-  FILE *file;
-  const char *value = "\\_SB.PCI0.LPC0.EC0.VPC0.DYTC 0x0013B001";
-
-  // Open /proc/acpi/call for writing with elevated privileges
-  file = fopen("/proc/acpi/call", "w");
-  if (file == NULL) {
-      perror("Error opening /proc/acpi/call");
-      return;
-  }
-
-  // Write the value to the file
-  if (fprintf(file, "%s", value) < 0) {
-      perror("Error writing to /proc/acpi/call");
-      fclose(file);
-      return;
-  }
-
-  // Close the file
-  fclose(file);
-  update_PwrMode_text(GTK_LABEL(label));
-  (void)widget;
-  (void)data;
-}
-
-static int is_rc_on(void)
-{
-    const char *acpi_mode;
-
-    acpi_mode = query_acpi_info("\\_SB.PCI0.LPC0.EC0.QCHO");
-    if (strncmp(acpi_mode, "0x0", 4))
-        return (1);
-    return (0);
-}
-
-static int is_bc_on(void)
-{
-    const char *acpi_mode;
-
-    acpi_mode = query_acpi_info("\\_SB.PCI0.LPC0.EC0.BTSM");
-    if (strncmp(acpi_mode, "0x0", 4))
-        return (1);
-    return (0);
-}
-
-static void write_acpi(const char *value)
-{
-  FILE *file;
-  file = fopen("/proc/acpi/call", "w");
-  if (file == NULL) {
-      perror("Error opening /proc/acpi/call");
-      return;
-  }
-  // Write the value to the file
-  if (fprintf(file, "%s", value) < 0) {
-      perror("Error writing to /proc/acpi/call");
-      fclose(file);
-      return;
-  }
-
-  // Close the file
-  fclose(file);
-}
-    
-static void toggle_battery_conservation(int mode)
-{
-    if (mode == 0)
-    {
-        printf("turning off battery conservation\n");
-        write_acpi("\\_SB.PCI0.LPC0.EC0.VPC0.SBMC 0x05");
-    }
-    else if (mode == 1)
-    {
-        printf("turning on battery conservation\n");
-        write_acpi("\\_SB.PCI0.LPC0.EC0.VPC0.SBMC 0x03");
-    }
-}
-
-static void toggle_rapid_charge(int mode)
-{
-    if (mode == 0)
-    {
-        printf("turning off rapid charge\n");
-        write_acpi("\\_SB.PCI0.LPC0.EC0.VPC0.SBMC 0x08");
-    }
-    else if (mode == 1)
-    {
-        printf("turning on rapid charge\n");
-        write_acpi("\\_SB.PCI0.LPC0.EC0.VPC0.SBMC 0x07");
-    }
-}
-static void switch_battery_bc(GtkWidget *widget, gpointer data)
-{
-  g_print ("Battery Conservation toggled\n");
-
-  if (is_bc_on())
-  {
-      toggle_battery_conservation(0);
-      gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(widget), is_bc_on());
-      return ;
-    //if both are on only shut down the one being toggled ie bc here
-  }
-  else if (is_rc_on())
-  {
-      toggle_rapid_charge(0);
-      gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(data), is_rc_on());
-      // otherwise we turn it on.
-  }
-  toggle_battery_conservation(1);
-  gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(widget), is_bc_on());
-}
-
-static void switch_battery_rc(GtkWidget *widget, gpointer data)
-{
-  g_print ("Rapid Charge toggled\n");
-
-  if (is_rc_on())
-  {
-      toggle_rapid_charge(0);
-      gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(widget), is_rc_on());
-      return ;
-    //if both are on only shut down the one being toggled ie bc here
-  }
-  else if (is_bc_on())
-  {
-      toggle_battery_conservation(0);
-      gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(data), is_bc_on());
-      // otherwise we turn it on.
-  }
-  toggle_rapid_charge(1);
-  gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(widget), is_rc_on());
-}
-
-static void
-quit_cb (GtkWindow *window)
-{
-  gtk_window_close (window);
-}
-
-static void
-activate (GtkApplication *app,
-          gpointer        user_data)
+static void activate (GtkApplication *app, gpointer user_data)
 {
   (void)user_data;
   /* Construct a GtkBuilder instance and load our UI description */
@@ -284,13 +14,13 @@ activate (GtkApplication *app,
   gtk_window_set_application (GTK_WINDOW (window), app);
 
   GObject *button = gtk_builder_get_object (builder, "btnIntelligentCooling");
-  g_signal_connect (button, "clicked", G_CALLBACK (set_IC), label);
+  g_signal_connect (button, "clicked", G_CALLBACK (set_PwrMode_IC), label);
 
   button = gtk_builder_get_object (builder, "btnExtremePerformance");
-  g_signal_connect (button, "clicked", G_CALLBACK (set_EP), label);
+  g_signal_connect (button, "clicked", G_CALLBACK (set_PwrMode_EP), label);
 
   button = gtk_builder_get_object (builder, "btnBatterySaving");
-  g_signal_connect (button, "clicked", G_CALLBACK (set_BS), label);
+  g_signal_connect (button, "clicked", G_CALLBACK (set_PwrMode_BS), label);
 
   update_PwrMode_text(GTK_LABEL(label));
   GObject *buttonBC = gtk_builder_get_object (builder, "btnBatteryBC");
@@ -299,17 +29,11 @@ activate (GtkApplication *app,
   gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(buttonRC), is_rc_on());
   g_signal_connect (buttonBC, "clicked", G_CALLBACK (switch_battery_bc), buttonRC);
   g_signal_connect (buttonRC, "clicked", G_CALLBACK (switch_battery_rc), buttonBC);
-  //button = gtk_builder_get_object (builder, "quit");
-  //g_signal_connect_swapped (button, "clicked", G_CALLBACK (quit_cb), window);
-  (void)quit_cb;
   gtk_widget_set_visible (GTK_WIDGET (window), TRUE);
-
-  /* We do not need the builder any more */
   g_object_unref (builder);
 }
 
-int
-main (int   argc, char *argv[])
+int main (int   argc, char *argv[])
 {
 #ifdef GTK_SRCDIR
   g_chdir (GTK_SRCDIR);
